perf: switch to using a web worker for saving a save file
